✨ Added >= operator to match helper
no issue

This feature adds a new operator to the match handlebars helper. It makes it possible to do a >= comparison which is usefull when you want to compare a price of a product.

Includes unit tests to check the new behaviour.
Run via `yarn test test/unit/frontend/helpers/match.test.js`